Move `DecCBOR` instances that have equivalent `Annotator`ones to testlib
except for `Raw` types wrapped in MemoBytes: `ShelleyTxAuxDataRaw`,
`ShelleyTxBodyRaw`, `AllegraTxBodyRaw`, `AlonzoTxBodyRaw`,
`BabbageTxBodyRaw`, `ConwayTxBodyRaw`
