+++ eras/allegra/impl/cardano-ledger-allegra.cabal
@@ -113,6 +113,7 @@ library testlib
     base,
     bytestring,
     cardano-ledger-allegra,
+    cardano-ledger-binary,
     cardano-ledger-core:{cardano-ledger-core, testlib},
     cardano-ledger-shelley:{cardano-ledger-shelley, testlib},
     cardano-strict-containers,
