+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Tx.hs
@@ -102,7 +102,6 @@ import Cardano.Ledger.Binary (
   Encoding,
   ToCBOR (..),
   decodeNullMaybe,
-  decodeNullStrictMaybe,
   encodeListLen,
   encodeNullMaybe,
   serialize,
@@ -406,23 +405,6 @@ instance
   ToCBOR (AlonzoTx era)
   where
   toCBOR = toEraCBOR @era
-
-instance
-  ( Typeable era
-  , DecCBOR (TxBody era)
-  , DecCBOR (TxWits era)
-  , DecCBOR (TxAuxData era)
-  ) =>
-  DecCBOR (AlonzoTx era)
-  where
-  decCBOR =
-    decode $
-      RecD AlonzoTx
-        <! From
-        <! From
-        <! From
-        <! D (decodeNullStrictMaybe decCBOR)
-  {-# INLINE decCBOR #-}
 
 instance
   ( Typeable era
