+++ eras/babbage/impl/cardano-ledger-babbage.cabal
@@ -76,6 +76,7 @@ library
   build-depends:
     aeson >=2.2,
     base >=4.18 && <5,
+    bytestring,
     cardano-data >=1.2,
     cardano-ledger-allegra ^>=1.8,
     cardano-ledger-alonzo ^>=1.14,
