+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Binary/Annotator.hs
@@ -1,9 +1,20 @@
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MonoLocalBinds #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Conway.Binary.Annotator where
+module Test.Cardano.Ledger.Conway.Binary.Annotator (
+  module Test.Cardano.Ledger.Babbage.Binary.Annotator,
+) where
+
+import Cardano.Ledger.Binary
+import Cardano.Ledger.Conway (ConwayEra)
+import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Conway.TxBody
+import Test.Cardano.Ledger.Babbage.Binary.Annotator
+
+deriving newtype instance DecCBOR (TxBody ConwayEra)
