+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Binary/Regression.hs
@@ -13,8 +13,9 @@ module Test.Cardano.Ledger.Conway.Binary.Regression where
 import Cardano.Ledger.BaseTypes (Inject (..), StrictMaybe (..), TxIx (..))
 import Cardano.Ledger.Binary (
   EncCBOR (..),
+  decCBOR,
   decodeFull,
-  decodeFullFromHexText,
+  decodeFullAnnotatorFromHexText,
   mkVersion,
   serialize,
  )
@@ -43,7 +44,7 @@ spec ::
 spec = describe "Regression" $ do
   it "DeserialiseFailure on resubmitting Conway Tx with invalid plutus script #4198" $ do
     io . expectRightDeep_ $
-      decodeFullFromHexText @(Tx era) (eraProtVerLow @era) $
+      decodeFullAnnotatorFromHexText @(Tx era) (eraProtVerLow @era) "Unwitnessed Tx" decCBOR $
         mconcat
           [ "84a700d9010282825820745f04573e7429be1404f9b936d208b81159f3fc4b300"
           , "37b9d630187eec1875600825820745f04573e7429be1404f9b936d208b81159f3"
@@ -58,7 +59,7 @@ spec = describe "Regression" $ do
           , "799f182aff0581840000d8799f182aff820000f4f6"
           ]
     expectRightDeep_ $
-      decodeFullFromHexText @(Tx era) (eraProtVerLow @era) $
+      decodeFullAnnotatorFromHexText @(Tx era) (eraProtVerLow @era) "Witnessed Tx" decCBOR $
         mconcat
           [ "84a700d9010282825820745f04573e7429be1404f9b936d208b81159f3fc4b300"
           , "37b9d630187eec1875600825820745f04573e7429be1404f9b936d208b81159f3"
