+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Binary/RoundTrip.hs
@@ -54,11 +54,11 @@ roundTripConwayCommonSpec ::
   , Arbitrary (PParamsHKD StrictMaybe era)
   , Arbitrary (InstantStake era)
   , Arbitrary (CertState era)
-  , DecCBOR (Annotator (Script era))
-  , DecCBOR (Annotator (TxAuxData era))
-  , DecCBOR (Annotator (TxWits era))
-  , DecCBOR (Annotator (TxBody era))
-  , DecCBOR (Annotator (Tx era))
+  , DecCBOR (Script era)
+  , DecCBOR (TxAuxData era)
+  , DecCBOR (TxWits era)
+  , DecCBOR (TxBody era)
+  , DecCBOR (Tx era)
   , RuleListEra era
   ) =>
   Spec
