+++ eras/mary/impl/cardano-ledger-mary.cabal
@@ -125,7 +125,7 @@ library testlib
     cardano-crypto-class,
     cardano-data:testlib,
     cardano-ledger-allegra:{cardano-ledger-allegra, testlib},
-    cardano-ledger-binary:testlib,
+    cardano-ledger-binary:{cardano-ledger-binary, testlib},
     cardano-ledger-core:{cardano-ledger-core, testlib},
     cardano-ledger-mary,
     cardano-ledger-shelley:{cardano-ledger-shelley, testlib},
