+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Tx.hs
@@ -13,7 +13,6 @@ module Cardano.Ledger.Shelley.Tx (
   auxDataShelleyTxL,
   sizeShelleyTxF,
   wireSizeShelleyTxF,
-  segWitTx,
   segWitAnnTx,
   mkBasicShelleyTx,
   shelleyMinFeeTx,
