+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Binary/RoundTrip.hs
@@ -48,11 +48,11 @@ roundTripShelleyCommonSpec ::
   , RuleListEra era
   , EraCertState era
   , Arbitrary (CertState era)
-  , DecCBOR (Annotator (Script era))
-  , DecCBOR (Annotator (TxAuxData era))
-  , DecCBOR (Annotator (TxWits era))
-  , DecCBOR (Annotator (TxBody era))
-  , DecCBOR (Annotator (Tx era))
+  , DecCBOR (Script era)
+  , DecCBOR (TxAuxData era)
+  , DecCBOR (TxWits era)
+  , DecCBOR (TxBody era)
+  , DecCBOR (Tx era)
   ) =>
   Spec
 roundTripShelleyCommonSpec = do
