+++ libs/cardano-protocol-tpraos/cardano-protocol-tpraos.cabal
@@ -88,6 +88,7 @@ library testlib
     cardano-crypto-class >=2.1.1,
     cardano-ledger-babbage,
     cardano-ledger-binary:{cardano-ledger-binary, testlib},
+    cardano-ledger-conway:testlib,
     cardano-ledger-core:{cardano-ledger-core, testlib},
     cardano-ledger-shelley:testlib,
     cardano-protocol-tpraos,
