+++ libs/cardano-protocol-tpraos/test/Test/Cardano/Protocol/Binary/BinarySpec.hs
@@ -21,6 +21,7 @@ import Test.Cardano.Ledger.Alonzo.Arbitrary ()
 import Test.Cardano.Ledger.Binary (decoderEquivalenceProp, decoderEquivalenceSpec)
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Mary.Arbitrary ()
+import Test.Cardano.Protocol.Binary.Annotator ()
 import Test.Cardano.Protocol.TPraos.Arbitrary ()
 
 spec :: Spec
