+++ libs/cardano-protocol-tpraos/testlib/Test/Cardano/Protocol/Binary/RoundTrip.hs
@@ -11,6 +11,7 @@ import Cardano.Ledger.Block (Block)
 import Cardano.Ledger.Core
 import Data.Typeable
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Core.Binary.Annotator ()
 import Test.Cardano.Ledger.Core.Binary.RoundTrip
 import Test.Cardano.Protocol.Binary.Annotator ()
 import Test.Cardano.Protocol.TPraos.Arbitrary ()
@@ -24,6 +25,7 @@ roundTripBlockSpec ::
   , EncCBOR h
   , EraSegWits era
   , Arbitrary (Block h era)
+  , DecCBOR (TxSeq era)
   ) =>
   Spec
 roundTripBlockSpec =
